Make spec-test optional
The CSS files look different depending on the OS. This implies that we can't commit the CSS files in order to keep our tests OS-independent. Since we probably still want to test for spec changes, this test is optional and can be run manually. I've added a short manual to test/spec.test.js.