Make the `indentedSyntax` option a bit smarter.
Instead of forcing loader config to set `sass?indentedSyntax` to
compile whitespace significant Sass files, automatically set it if
necessary (if the current file is a SASS file).

This should make loader config a bit easier, because now you won’t need
to have separate loaders for SCSS And SASS.

Additionally this has the side benefit of enabling SCSS <-> SASS
interop in some advanced cases. Without this, it wasn’t possible for my
[sprockets-directive-loader](https://github.com/timmfin/sprockets-direct
ive-loader) project to work when a SASS file `//= require`ed a SCSS
file. That’s because you would have had to set `indentedSyntax` on the
loader, and that would have been “passed-on” (via
`css?importLoaders=1`) to the SCSS file, causing its compile to fail.

So with this change, this will now work for a mix of SCSS and SASS
files:

```
{
test: /\.(scss|sass)$/,
loader: ExtractTextPlugin.extract("style",
"css?sourceMap&importLoaders=2!sprockets-directive!sass?sourceMap")
}
```