moved node-sass to dependencies
Not sure if this is a good idea or not, but I had to do this in order to get webpack working locally so I figured I'd submit it. I was getting errors about missing a particular version of node-sass, so it seems like it should just be a req of this module.