docs(readme): Adding extract-css-chunks
Adding another option to the readme, it’s mini-css with some built-in capabilities to support hot module reloading of css files in development.

It’s super fast and more debuggable than having to use style-loader to append massive amounts of css. 

It would be a great addition for users looking to have dev mimic production closely, it should help cascading collisions and general performance in the browser as the dom won’t have huge css written into its head.